requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2.3
django-ckeditor-5==0.2.18
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
idna==3.10
pillow==11.2.1
requests==2.32.4
soupsieve==2.7
sqlparse==0.5.3
typing_extensions==4.14.0
urllib3==2.4.0
